[otbn] Move Python library code into a shared/ directory
