+++ hw/ip/otbn/util/otbn-as
@@ -24,8 +24,8 @@ import sys
 import tempfile
 from typing import Dict, List, Optional, Set, TextIO, Tuple
 
-from insn_yaml import (BitRanges, Encoding, Insn, InsnsFile, Operand,
-                       RegOperandType, load_file)
+from shared.insn_yaml import (BitRanges, Encoding, Insn, InsnsFile, Operand,
+                              RegOperandType, load_file)
 
 
 class RVFmt:
